I think we now pass all tests. Removed extra slater files. Now there's just one slater object that does everything!
