+++ tests/integration/test_ecp.py
@@ -1,8 +1,7 @@
 import pandas as pd
 from pyqmc.mc import vmc, initial_guess
 from pyscf import gto, scf
-from pyqmc.slater import PySCFSlater
-from pyqmc.multislater_orbs import MultiSlater
+from pyqmc import Slater
 from pyqmc.accumulators import EnergyAccumulator
 import numpy as np
 
@@ -12,7 +11,7 @@ def test_ecp():
     mol = gto.M(atom="C 0. 0. 0.", ecp="bfd", basis="bfd_vtz")
     mf = scf.RHF(mol).run()
     nconf = 5000
-    wf = MultiSlater(mol, mf)
+    wf = Slater(mol, mf)
     coords = initial_guess(mol, nconf)
     df, coords = vmc(
         wf, coords, nsteps=100, accumulators={"energy": EnergyAccumulator(mol)}
