+++ tests/integration/test_obdm.py
@@ -8,7 +8,7 @@ import numpy as np
 np.random.seed(12534234)
 from pyscf import gto, scf, lo
 from numpy.linalg import solve
-from pyqmc import MultiSlater
+from pyqmc import Slater
 from pyqmc.mc import initial_guess, vmc
 from pandas import DataFrame
 from pyqmc.obdm import OBDMAccumulator, normalize_obdm
@@ -34,7 +34,7 @@ def test():
     nconf = 500
     nsteps = 400
     warmup = 15
-    wf = MultiSlater(mol, mf)
+    wf = Slater(mol, mf)
     configs = initial_guess(mol, nconf)
     obdm_dict = dict(mol=mol, orb_coeff=lowdin, nsweeps=5, warmup=15)
     obdm = OBDMAccumulator(**obdm_dict)
@@ -100,7 +100,7 @@ def test_pbc():
     nconf = 800
     nsteps = 50
     warmup = 6
-    wf = MultiSlater(mol, mf)
+    wf = Slater(mol, mf)
     configs = initial_guess(mol, nconf)
     obdm_dict = dict(mol=mol, orb_coeff=lowdin, kpts=kpts, nsweeps=4, warmup=10)
     obdm = OBDMAccumulator(**obdm_dict)
