+++ tests/unit/test_derivatives.py
@@ -16,7 +16,7 @@ def test_wfs():
     """
 
     from pyscf import lib, gto, scf
-    from pyqmc.slater import PySCFSlater
+    from pyqmc import Slater
     from pyqmc.multiplywf import MultiplyWF
     from pyqmc.manybody_jastrow import J3
     from pyqmc import default_jastrow
@@ -32,11 +32,11 @@ def test_wfs():
     for wf in [
         default_jastrow(mol)[0],
         J3(mol),
-        MultiplyWF(PySCFSlater(mol, mf), default_jastrow(mol)[0]),
-        MultiplyWF(PySCFSlater(mol, mf), default_jastrow(mol)[0], J3(mol)),
-        PySCFSlater(mol, mf_uhf),
-        PySCFSlater(mol, mf),
-        PySCFSlater(mol, mf_rohf),
+        MultiplyWF(Slater(mol, mf), default_jastrow(mol)[0]),
+        MultiplyWF(Slater(mol, mf), default_jastrow(mol)[0], J3(mol)),
+        Slater(mol, mf_uhf),
+        Slater(mol, mf),
+        Slater(mol, mf_rohf),
     ]:
         for k in wf.parameters:
             if k != "mo_coeff":
@@ -57,9 +57,7 @@ def test_wfs():
                 testwf.test_wf_pgradient,
             ],
         ):
-            err = []
-            for delta in [1e-4, 1e-5, 1e-6, 1e-7, 1e-8]:
-                err.append(func(wf, epos, delta)[0])
+            err = [func(wf, epos, delta)[0] for delta in [1e-4, 1e-5, 1e-6, 1e-7, 1e-8]]
             print(type(wf), fname, min(err))
             assert min(err) < epsilon, "epsilon {0}".format(epsilon)
 
@@ -71,8 +69,7 @@ def test_pbc_wfs():
 
     from pyscf.pbc import lib, gto, scf
     from pyqmc.supercell import get_supercell
-    from pyqmc.slater import PySCFSlater
-    from pyqmc.multislaterpbc import MultiSlaterPBC
+    from pyqmc.slater import Slater
     from pyqmc.multiplywf import MultiplyWF
     from pyqmc import default_jastrow
     import pyqmc
@@ -91,19 +88,16 @@ def test_pbc_wfs():
     supercell = get_supercell(mol, S=(np.ones((3, 3)) - 2 * np.eye(3)))
     epos = pyqmc.initial_guess(supercell, nconf)
     # For multislaterpbc
-    kinds = 0, 3, 5, 6  # G, X, Y, Z
-    d1 = {kind: [0] for kind in kinds}
-    d2 = d1.copy()
-    d2.update({0: [], 3: [0, 1]})
-    detwt = [2 ** 0.5, 2 ** 0.5]
-    occup = [[d1, d2], [d1]]
-    map_dets = [[0, 1], [0, 0]]
+    #kinds = 0, 3, 5, 6  # G, X, Y, Z
+    #d1 = {kind: [0] for kind in kinds}
+    #d2 = d1.copy()
+    #d2.update({0: [], 3: [0, 1]})
+    #detwt = [2 ** 0.5, 2 ** 0.5]
+    #occup = [[d1, d2], [d1]]
+    #map_dets = [[0, 1], [0, 0]]
     for wf in [
-        MultiplyWF(PySCFSlater(supercell, mf), default_jastrow(supercell)[0]),
-        PySCFSlater(supercell, mf),
-        MultiSlaterPBC(supercell, mf, detwt=detwt, occup=occup, map_dets=map_dets),
-        # PySCFSlaterPBC(supercell, mf_uhf),
-        # PySCFSlaterPBC(supercell, mf_rohf),
+        MultiplyWF(Slater(supercell, mf), default_jastrow(supercell)[0]),
+        Slater(supercell, mf),
     ]:
         for k in wf.parameters:
             if "mo_coeff" not in k and k != "det_coeff":
@@ -119,9 +113,7 @@ def test_pbc_wfs():
                 testwf.test_wf_pgradient,
             ],
         ):
-            err = []
-            for delta in [1e-4, 1e-5, 1e-6, 1e-7, 1e-8]:
-                err.append(func(wf, epos, delta)[0])
+            err = [func(wf, epos, delta)[0] for delta in [1e-4, 1e-5, 1e-6, 1e-7, 1e-8]]
             print(type(wf), fname, min(err))
             assert min(err) < epsilon
