+++ tests/unit/test_overlap_derivative.py
@@ -66,7 +66,7 @@ def test():
     mf.stdout = None
 
     print("setting up wfs", flush=True)
-    wf0 = pyqmc.MultiSlater(mol, mf)
+    wf0 = pyqmc.Slater(mol, mf)
     mf.mo_coeff[0][:, 0] = np.mean(mf.mo_coeff[0][:, :2], axis=1)
     wf1, to_opt = pyqmc.default_slater(mol, mf, optimize_orbitals=True)
